+++ src/app/(route)/(monster)/monster/produce/components/WriteStory.tsx
@@ -5,8 +5,8 @@ import { Controller, useFormContext } from 'react-hook-form';
 
 import TextField from '@components/common/TextField';
 
-import { UserInfo } from '../../(auth)/signup/context/context.type';
-import useSignUpContext from '../../(auth)/signup/hooks/useSignUpContext';
+import { UserInfo } from '../../../../(auth)/signup/context/context.type';
+import useSignUpContext from '../../../../(auth)/signup/hooks/useSignUpContext';
 
 const MAX_LENGTH = 500;
